backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
python-multipart
python-jose[cryptography]
